Extracted NammuConsoleHandler class outside NammuController.
