+++ python/nammu/controller/NammuController.py
@@ -27,6 +27,7 @@ from ModelController import ModelController
 from ToolbarController import ToolbarController
 from ..SOAPClient.SOAPClient import SOAPClient
 from ..view.NammuView import NammuView
+from ..utils.NammuConsoleHandler import NammuConsoleHandler
 
 
 class NammuController(object):
@@ -649,24 +650,3 @@ class NammuController(object):
         logger.addHandler(console_handler)    
 
         return logger
-    
-    
-class NammuConsoleHandler(StreamHandler):
-    """
-    Extends StreamHandler to make it print log messages in Nammu's console for
-    the user to see.
-    """
-    def __init__(self, nammu_console):
-        """
-        Needs a reference to nammu to be able to get ahold of Nammu's console.
-        """
-        super(logging.StreamHandler, self).__init__()
-        self.nammu_console = nammu_console
-
-    def emit(self, record):
-        """
-        This is the method that prints out the log message. Format the given
-        record and send to Nammu's console for the user to see.
-        """
-        msg = self.format(record)
-        self.nammu_console.addText(msg + "\n")
